Derived simulation class example: SEIR percolation
Initial commit for an example derived class.  It's not really scientifically interesting, but it illustrates how one of the built-in classes can be used as the base class for a new, similar type of simulation that the user might want to do.

The tmp_research program with the SEIR_percolation class perform a small simulation with a cute visualization tool to track who is getting infected when.  It sometimes is apparent that certain nodes are rarely infected because their degree is low, and other times because they are in an isolated component/region of the network.