Major ChainBin bug fix, many smaller improvements.
ChainBinomial_Sim had a bug that stochastically resulted in seg faults by trying to access a node that didn't exist.

Network::read_edgelist() was overly sensitive to whitespace and would treat imported node names as different if there was leading or trailing whitespace.  write_edgelist() always wrote to the same file before; also fixed.

Simulation derived classes are a bit safer, and the base class is a bit more useful with the addition of a few more virtual functions.

strip() function added to Utility.cpp, which can remove arbitrary leading or trailing characters from a string.