Many chances to how axes are labeled and ticked.
One major bug fix: the ranges of axes were sometimes being extended in order
to put tick marks in nice positions, but the scene coordinates were not being
updated.  This resulted in plotted data not being rescaled to reflect the fact
that the axis range was bigger than the data range.

The plotAxis interface isn't exactly elegant, but it is more powerful.  Axes may
have a prefered or required number of ticks, and if intLabels is true, paint()
will try to accommodate.  Tick marks are prefered at multiples of 1, 2, 3, 5 or
10, scaled to a sensible order of magnitude given the range specified, such that
no more than 10 ticks will be drawn.

Tick labels are now centered on ticks, and if there's not enough room to show
every label, we print every other one.  This makes the plots much cleaner in
appearance, through a large range of window sizes.