New status indicators (cursor, status bar)
New status bar provides some guidance regarding what should be clicked when,
and indicates when something time consuming is going on in the background;
the cursor also now will change to the wait cursor.

Now we also calculate a guess of what the epidemic size will be, based on
the SIR mass-action model (if one fails to occur stochastically, this guess
will be way off).  This should allow us to give the user a progress report
during simulations.  Something similar for network construction should be
possible, but I'm not sure how it should be done yet.  We certainly know
the expected number of nodes and edges, so it's just a matter of communication.