Improvements to poisson random graph generation
Added fast_random_graph(), which tries to pick the best algorithm, depending on the probability that any two nodes will be connected.  Fixed some precision/overflow issues related to generating poisson distributions: for lambda > 500 we use a normal approximation of the poisson, which avoids evaluating huge factorials and exponents.  For lambda < 500, we use long doubles now; previously, overflow occurred when lambda was around 80.  Added normal_cdf() and normal_pdf() functions for use in the approximations.