Fixed net component bug, adding process control
There was a bug in get_major_component() (now called get_biggest_component()) that made it likely to fail unless the first node in the node list was in a component that was more than half the network.  I've added a get_components() function and made the code a bit more generic.

It has become clear that in order to cancel long-running processes like network generation from within the EpiFire gui, we need some way of communicating with the network libraries.  The least invasive way I could think of doing this was to add a process_stopped attribute that could be set from the outside and checked on the inside of the network code.  This implementation is working but there's more to be done.