Shortest path functions now use single component
One bug fix: when calculating mean path length, denominator needed to be reduced by one because we don't consider the distance from a node to itself.

Other (more generally consequential) changes: the functions that directly or indirectly call min_path() (i.e., Dijkstra's algorithm) act on only a subset of nodes provided as an argument.  If the user provides an empty subset, than the answer for the entier network is calculated.  Keep in mind that this may result in some nodes being an infinite distance apart, which in this case is denoted by a distance of -1.