More multithreading, better user control
Most of the changes have focused on moving the simulations into a second thread and on improving the user's ability to cancel threads in progress.  progressDialog is no longer automatically shown and hidden whenever threads started and stopped (that was just annoying).  We're now waiting/sleeping after threads are cancelled; this fixed a number of issues caused by thread collisions if users tried to do certain things before resources were freed up.  The degree distribution plots no longer look so damn ugly--it was a problem with how the number of bins was selected.  Temporary changes to saveData(), which still is not functional.