Fixed edge count; sim status updates use signals
Previously the edge count in the network analysis was the number of outbound
edges in the network, which is not how most people would think of edges in
an undirected network.  Now that number is divided in two if it's undirected.

Also, the logEditor was being updated directly from within the background
thread, which could sometimes cause crashes.  Now updates are being done via
a signal.