Changed split() behavior, added parameter validation
The most significant change here is that split() now returns the whole string
if no delimiters are present.  This is useful for parsing csv files, in case
a given line has only 1 value with no comma.

gen_trunc_exponential() now verifies that the parameter it was given is positive,
since zero and negatives are undefined.

Fixed inaccurate error message.