Retained data bugfix; added makeEditable()
Although data was not internally being retained if retain data was
unchecked, the epi curve plot was not being cleared, so it started
getting ugly.

Since it doesn't make sense for number of nodes to be editable for
imported networks, it no longer is.  Added makeEditable() to facil-
itate switching back and forth.