Refined behavior during intensive processes
In OS X, the progress dialog was always attached to the main window,
even when the process that triggered it was coming from the net
analysis dialog.  I've made the progress dialog parentless, and this
seems to have fixed the issue.

In order to prevent users from spawning multiple simulation threads
(not thread-safe, since states in the same network are being written),
we were ignoring clicks; since the progress dialog is also in the main
thread, however, click events were also being ignored for the cancel
button.  The best solution would likely be to implement an event handler.
For the time being, however, I'm disabling the simulation and network
buttons while the simulation is running, and disabling the entire central
widget if the progress dialog pops up.

The cursor was not always returning to an arrow when it should after running
the network analysis dialog.  This is fixed.