Adding library config file for OSX
Run this shell script In order to compile the osx version using distributed (rather than global)
copies of the Qt libraries.  This makes the install file quite a bit bigger than it would be
otherwise, but this way users don't have to install the much larger Qt libraries.  Also,
dependency problems for end-users are less likely if they're already getting the right
versions of the Qt libraries.