Histogram bugfix; results analysis has histogram
Data were not always being put into the correct bins; this was particularly
evident when there were few bins and for data at the upper end of the range.
This is now fixed.

The results analysis now has its own histogram that allows the user to change
the range and the number of bins.  It's now much more useful for data exploration.