Two minor Network bugfixes
If Edge::break_end() is called but the end is already NULL (broken), this
previously caused a crash (fixed).

if (Network::rand_connect_stubs() was called with an empty stubs vector, program
would crash when it tried to shuffle the empty vector. Function now returns if
vector is empty (fixed).

I'm also deleting the histogram.* files, which we are no longer using.