Fixed 2 minor mainWindow bugs
Expected R-zero value and mass-action prediction were not updating when
new networks were being generated because the generating degree distribution
rather than the actual degree distribution were being used to calculate the
critical transmissibility.  Not a huge problem, but probably not what the user
was expecting.  This did not affect Poisson network generation, because that
was not using the Molloy-Reed algorithm, and therefore the generating degree
distribution was never pre-defined.  Anyway, it's fixed.

The parent of one of the LineEdit validators was wrong.  Doesn't appear to have
caused abberant behavior, but I fixed it.