normalize_dist() bugfix affecting epi predictions
normalize_dist() was ignoring the index 0 value, which was reasonable
if that value was always zero.  Which a long time ago, was a reasonable
assumption given the use of that function.  This function is usually
called on degree distributions, and many networks don't have degree
zero nodes.  If they do, however, the resulting normalized distribution
is wrong, which meant that all theoretical predictions that depend on
the degree distribution would also be wrong.

In the EpiFire GUI, this bug manifested itself with theoretical epi size
predictions (network model) that were too large when Poisson graphs were
used, or potentially when users imported a network.

Oh, and normalize_dist() is a template function now.  A bit cleaner.