Net analysis on imported net bugfix
Epifire was complaining that a network did not exist when users tried
to analyze the network after importing.  It's a bug with how the network
pointer gets updated, but I'm not sure why the same thing doesn't happen
for generated networks.  In any case, we now force the pointer to update,
so now the behavior is fixed.