Many changes to plot classes
Short version: All plots have axes and axis labels, and plots now comprise
functional regions (internally).  A few minor bugs got fixed.

Long version:  The plot itself (where the dots/rectangles are) is a
PlotRegion and it is surrounded on all four sides by axis areas, which
are in turn surrounded by margin areas, all of which are also PlotRegions.
This allows us to use a convenient coordinate system for drawing things inside,
since each PlotRegion has its own coordinates, which makes aligning things much
more straightforward.

The node state plot is a bit more clever now, in that it stays lined up
with the red epi curve.  The exception is when the epi curve spans its x-axis,
which is a problem because the epi curve points are centered on the ticks, but
the node state rectangles ("pixels") are left-justified.  In that case, we make
the node state axis a bit longer, so that the plot does not become misleading.

Two minor bugs got fixed.  In the results analysis dialog, if users set a
maximum range value and then deleted it, it defaulted to 0, which meant that
the range was [0,0].  Now it defaults to the nonsensical -1, which means that
the max data value should be used.

Also, if the max x value in the epi curve plot was less than 9, 10 ticks were
still used, which meant that there were more ticks than integers, which meant
that some labels were repeated, which was stupid.