Implemented new ChainBinomial algorithm
At the MIDAS meeting in Pittsburgh (Oct 2010), Joel Miller pointed out that there
was an alternate approach to the chain binomial algorithm that is lower complexity.
Basically, instead of testing each neighbor of an infectious node d times, where
d is the infectious period, we generate one random nonuniform integer that determines
whether the neighbor gets infected and on what day.

In practice, the two approaches are about the same if infectious period is small (the
old method can even be faster if it's very small), but as infectious period gets large
(say, >15) the new algorithm becomes increasingly advantageous.