Tweaking buttons & menus; minor bugfix
Adding draw network button, removing reset to defaults button.

Bugfix: if network plot was open and a network that is too large is
generated, the old network was retained in plot, but simulations happened
on new network.  Simulations nonsensically were then mapped to old network.
Network plot is now cleared if the network is too large to plot (currently
max 10,000 nodes).