Resolved segfaults on gui updates from 2nd thread
Because of changes to how multithreading is handled outside of the scope of
EpiFire, some of the things that used to work no longer do.  Specifically, we
used to be able to make updates to the GUI from a background thread, and that
no longer works.  We're now storing data from the background thread and then
checking whether it's been updated after the thread closes, rather than trying
to do updates directly.