Mean path length bugfix
We were correctly calculating pairwise path lengths, but not including all
of them when averaging to get the mean distance between nodes in the network.
The mean_dist() function incorrectly assumed that cells with the same row and
column index in the distance matrix referred to a node's distance to itself,
but the data structure only contained the upper triangle of the matrix.